feat: add ip pool name in integration
