+++ apps/web/src/pages/integrations/IntegrationsStorePage.tsx
@@ -12,6 +12,7 @@ import {
   InAppProviderIdEnum,
   ProvidersIdEnum,
   SmsProviderIdEnum,
+  ICredentials,
 } from '@novu/shared';
 
 import PageHeader from '../../components/layout/components/PageHeader';
@@ -150,34 +151,6 @@ export interface IIntegratedProvider {
   logoFileName: ILogoFileName;
   betaVersion: boolean;
   novu?: boolean;
-}
-
-export interface ICredentials {
-  apiKey?: string;
-  user?: string;
-  secretKey?: string;
-  domain?: string;
-  password?: string;
-  host?: string;
-  port?: string;
-  secure?: boolean;
-  region?: string;
-  accountSid?: string;
-  messageProfileId?: string;
-  token?: string;
-  from?: string;
-  senderName?: string;
-  applicationId?: string;
-  clientId?: string;
-  projectName?: string;
-  serviceAccount?: string;
-  baseUrl?: string;
-  webhookUrl?: string;
-  requireTls?: boolean;
-  ignoreTls?: boolean;
-  tlsOptions?: Record<string, unknown>;
-  redirectUrl?: string;
-  hmac?: boolean;
 }
 
 export interface IntegrationEntity {
