+++ apps/web/src/pages/integrations/components/ConnectIntegrationForm.tsx
@@ -11,7 +11,7 @@ import { ChannelTypeEnum, ICredentialsDto, IConfigCredentials } from '@novu/shar
 import { Button, colors, Input, Switch, Text } from '../../../design-system';
 import { IIntegratedProvider } from '../IntegrationsStorePage';
 import { createIntegration, getWebhookSupportStatus, updateIntegration } from '../../../api/integration';
-import { Close } from '../../../design-system/icons/actions/Close';
+import { Close } from '../../../design-system/icons';
 import { IntegrationInput } from './IntegrationInput';
 import { IS_DOCKER_HOSTED, WEBHOOK_URL } from '../../../config';
 import { useEnvController, useAuthController } from '../../../hooks';
