+++ libs/dal/src/repositories/integration/integration.entity.ts
@@ -1,37 +1,8 @@
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 
 import type { EnvironmentId } from '../environment';
 import type { OrganizationId } from '../organization';
 import { ChangePropsValueType } from '../../types/helpers';
-
-export interface ICredentials {
-  apiKey?: string;
-  user?: string;
-  secretKey?: string;
-  domain?: string;
-  password?: string;
-  host?: string;
-  port?: string;
-  secure?: boolean;
-  region?: string;
-  accountSid?: string;
-  messageProfileId?: string;
-  token?: string;
-  from?: string;
-  senderName?: string;
-  applicationId?: string;
-  clientId?: string;
-  projectName?: string;
-  serviceAccount?: string;
-  baseUrl?: string;
-  webhookUrl?: string;
-  ipPoolName?: string;
-  requireTls?: boolean;
-  ignoreTls?: boolean;
-  tlsOptions?: Record<string, unknown>;
-  redirectUrl?: string;
-  hmac?: boolean;
-}
 
 export class IntegrationEntity {
   _id: string;
@@ -44,7 +15,7 @@ export class IntegrationEntity {
 
   channel: ChannelTypeEnum;
 
-  credentials: ICredentials;
+  credentials: ICredentialsEntity;
 
   active: boolean;
 
@@ -58,5 +29,7 @@ export class IntegrationEntity {
 
   deletedBy: string;
 }
+
+export type ICredentialsEntity = ICredentials;
 
 export type IntegrationDBModel = ChangePropsValueType<IntegrationEntity, '_environmentId' | '_organizationId'>;
