+++ libs/shared/src/consts/providers/provider.enum.ts
@@ -26,6 +26,7 @@ export enum CredentialsKeyEnum {
   TlsOptions = 'tlsOptions',
   RedirectUrl = 'redirectUrl',
   Hmac = 'hmac',
+  IpPoolName = 'ipPoolName',
 }
 
 export enum EmailProviderIdEnum {
