+++ libs/shared/src/dto/integration/construct-integration.interface.ts
@@ -1,30 +1,6 @@
-export interface ICredentialsDto {
-  apiKey?: string;
-  user?: string;
-  secretKey?: string;
-  domain?: string;
-  password?: string;
-  host?: string;
-  port?: string;
-  secure?: boolean;
-  hmac?: boolean;
-  region?: string;
-  accountSid?: string;
-  messageProfileId?: string;
-  token?: string;
-  from?: string;
-  senderName?: string;
-  applicationId?: string;
-  clientId?: string;
-  projectName?: string;
-  serviceAccount?: string;
-  baseUrl?: string;
-  requireTls?: boolean;
-  ignoreTls?: boolean;
-  tlsOptions?: Record<string, unknown>;
-  webhookUrl?: string;
-  redirectUrl?: string;
-}
+import { ICredentials } from '../../entities/integration';
+
+export type ICredentialsDto = ICredentials;
 
 export interface IConstructIntegrationDto {
   credentials: ICredentialsDto;
