+++ libs/shared/src/entities/integration/index.ts
@@ -1,0 +1,1 @@
+export * from './credential.interface';
