+++ packages/application-generic/src/factories/chat/handlers/mattermost.handler.ts
@@ -1,4 +1,4 @@
-import { ICredentials } from '@novu/dal';
+import { ICredentials } from '@novu/shared';
 import { ChannelTypeEnum } from '@novu/stateless';
 import { BaseChatHandler } from './base.handler';
 import { MattermostProvider } from '@novu/mattermost';
