+++ packages/application-generic/src/factories/chat/handlers/slack.handler.ts
@@ -1,4 +1,4 @@
-import { ICredentials } from '@novu/dal';
+import { ICredentials } from '@novu/shared';
 import { ChannelTypeEnum } from '@novu/stateless';
 import { SlackProvider } from '@novu/slack';
