+++ packages/application-generic/src/factories/chat/interfaces/index.ts
@@ -1,6 +1,6 @@
 import { IChatOptions, ISendMessageSuccessResponse } from '@novu/stateless';
-import { ICredentials, IntegrationEntity } from '@novu/dal';
-import { ChannelTypeEnum } from '@novu/shared';
+import { IntegrationEntity } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 
 export interface IChatHandler {
   canHandle(providerId: string, channelType: ChannelTypeEnum);
