+++ packages/application-generic/src/factories/mail/handlers/email-webhook.handler.ts
@@ -1,6 +1,5 @@
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { EmailWebhookProvider } from '@novu/email-webhook';
-import { ICredentials } from '@novu/dal';
 import { BaseHandler } from './base.handler';
 export class EmailWebhookHandler extends BaseHandler {
   constructor() {
