+++ packages/application-generic/src/factories/mail/handlers/emailjs.handler.ts
@@ -1,7 +1,6 @@
 import { IEmailJsConfig } from '@novu/emailjs/build/main/lib/emailjs.config';
 import { EmailJsProvider } from '@novu/emailjs';
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { BaseHandler } from './base.handler';
 
 /**
