+++ packages/application-generic/src/factories/mail/handlers/mailersend.handler.ts
@@ -1,5 +1,4 @@
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { MailersendEmailProvider } from '@novu/mailersend';
 
 import { BaseHandler } from './base.handler';
