+++ packages/application-generic/src/factories/mail/handlers/mailgun.handler.ts
@@ -1,6 +1,5 @@
 import { MailgunEmailProvider } from '@novu/mailgun';
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { BaseHandler } from './base.handler';
 
 export class MailgunHandler extends BaseHandler {
