+++ packages/application-generic/src/factories/mail/handlers/postmark.handler.ts
@@ -1,5 +1,4 @@
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { PostmarkEmailProvider } from '@novu/postmark';
 import { BaseHandler } from './base.handler';
