+++ packages/application-generic/src/factories/mail/handlers/resend.handler.ts
@@ -1,5 +1,4 @@
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { ResendEmailProvider } from '@novu/resend';
 import { BaseHandler } from './base.handler';
