+++ packages/application-generic/src/factories/mail/handlers/ses.handler.ts
@@ -1,5 +1,4 @@
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { SESConfig } from '@novu/ses/build/module/lib/ses.config';
 import { SESEmailProvider } from '@novu/ses';
 import { BaseHandler } from './base.handler';
