+++ packages/application-generic/src/factories/mail/interfaces/send.handler.interface.ts
@@ -4,8 +4,7 @@ import {
   ISendMessageSuccessResponse,
   ICheckIntegrationResponse,
 } from '@novu/stateless';
-import { ICredentials } from '@novu/dal';
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 
 export interface IMailHandler {
   canHandle(providerId: string, channelType: ChannelTypeEnum);
