+++ packages/application-generic/src/factories/push/handlers/apns.handler.ts
@@ -1,7 +1,6 @@
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { APNSPushProvider } from '@novu/apns';
 import { BasePushHandler } from './base.handler';
-import { ICredentials } from '@novu/dal';
 
 export class APNSHandler extends BasePushHandler {
   constructor() {
