+++ packages/application-generic/src/factories/push/handlers/fcm.handler.ts
@@ -1,7 +1,6 @@
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { FcmPushProvider } from '@novu/fcm';
 import { BasePushHandler } from './base.handler';
-import { ICredentials } from '@novu/dal';
 
 export class FCMHandler extends BasePushHandler {
   constructor() {
