+++ packages/application-generic/src/factories/push/interfaces/push.handler.interface.ts
@@ -1,6 +1,5 @@
 import { IPushOptions, ISendMessageSuccessResponse } from '@novu/stateless';
-import { ICredentials } from '@novu/dal';
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 
 export interface IPushHandler {
   canHandle(providerId: string, channelType: ChannelTypeEnum);
