+++ packages/application-generic/src/factories/sms/handlers/firetext.handler.ts
@@ -1,5 +1,4 @@
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { BaseSmsHandler } from './base.handler';
 import { FiretextSmsProvider } from '@novu/firetext';
