+++ packages/application-generic/src/factories/sms/handlers/infobip.handler.ts
@@ -1,6 +1,5 @@
 import { InfobipSmsProvider } from '@novu/infobip';
-import { ChannelTypeEnum, SmsProviderIdEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, SmsProviderIdEnum, ICredentials } from '@novu/shared';
 import { BaseSmsHandler } from './base.handler';
 export class InfobipSmsHandler extends BaseSmsHandler {
   constructor() {
