+++ packages/application-generic/src/factories/sms/handlers/kannel.handler.ts
@@ -1,6 +1,5 @@
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { KannelSmsProvider } from '@novu/kannel';
-import { ICredentials } from '@novu/dal';
 import { BaseSmsHandler } from './base.handler';
 
 export class KannelSmsHandler extends BaseSmsHandler {
