+++ packages/application-generic/src/factories/sms/handlers/maqsam.handler.ts
@@ -1,6 +1,5 @@
 import { MaqsamSmsProvider } from '@novu/maqsam';
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { BaseSmsHandler } from './base.handler';
 
 export class MaqsamHandler extends BaseSmsHandler {
