+++ packages/application-generic/src/factories/sms/handlers/plivo.handler.ts
@@ -1,6 +1,5 @@
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { PlivoSmsProvider } from '@novu/plivo';
-import { ICredentials } from '@novu/dal';
 import { BaseSmsHandler } from './base.handler';
 
 export class PlivoHandler extends BaseSmsHandler {
