+++ packages/application-generic/src/factories/sms/handlers/sms-central.handler.ts
@@ -1,5 +1,4 @@
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { BaseSmsHandler } from './base.handler';
 import { SmsCentralSmsProvider } from '@novu/sms-central';
