+++ packages/application-generic/src/factories/sms/handlers/sns.handler.ts
@@ -1,7 +1,6 @@
-import { ChannelTypeEnum } from '@novu/shared';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { SNSSmsProvider } from '@novu/sns';
 import { SNSConfig } from '@novu/sns/build/main/lib/sns.config';
-import { ICredentials } from '@novu/dal';
 import { BaseSmsHandler } from './base.handler';
 
 export class SnsHandler extends BaseSmsHandler {
