+++ packages/application-generic/src/factories/sms/handlers/twilio.handler.ts
@@ -1,6 +1,5 @@
 import { TwilioSmsProvider } from '@novu/twilio';
-import { ChannelTypeEnum } from '@novu/shared';
-import { ICredentials } from '@novu/dal';
+import { ChannelTypeEnum, ICredentials } from '@novu/shared';
 import { BaseSmsHandler } from './base.handler';
 
 export class TwilioHandler extends BaseSmsHandler {
